requirements.txt
openai-whisper
pydub
librosa
click
scipy
numpy
PyAudio
keyboard
torch